some more dumb crap

+++ src/main/java/frc/robot/PositionPreset.java
@@ -1,20 +1,20 @@
 package frc.robot;
 
 public enum PositionPreset {
-  HUMAN_PLAYER_PICKUP_CONE(90.0, 25.0),
-  HUMAN_PLAYER_PICKUP_CUBE(106.0, 25.0),
-  FLOOR_CONE(0.0, 75.0),
-  FLOOR_CUBE(0.0, 80.0),
-  HIGH_CONE(105.0, 27.0),
-  MID_CONE(109.0, 20.0),
-  CUBE(90, 40.0),
-  STARTING(0.0, 0.0),
-  DRIVING(60.0, 0.0);
+  HUMAN_PLAYER_PICKUP_CONE(90.0F, 25.0F),
+  HUMAN_PLAYER_PICKUP_CUBE(106.0F, 25.0F),
+  FLOOR_CONE(0.0F, 75.0F),
+  FLOOR_CUBE(0.0F, 80.0F),
+  HIGH_CONE(105.0F, 27.0F),
+  MID_CONE(109.0F, 20.0F),
+  CUBE(90F, 40.0F),
+  STARTING(0.0F, 0.0F),
+  DRIVING(60.0F, 0.0F);
   
-  public final double upperPivotAngle;
-  public final double winchAngle;
+  public final float upperPivotAngle;
+  public final float winchAngle;
 
-  PositionPreset(double upperPivotAngle, double winchAngle) {
+  PositionPreset(float upperPivotAngle, float winchAngle) {
     this.upperPivotAngle = upperPivotAngle;
     this.winchAngle = winchAngle;
   }
